Indent commands like >>, << and <C-T>/<C-D> in insert mode indent the entire line
For some kinds of comments, like the big boilerplate at the file header etc., the comment prefix (e.g. # for Perl scripts) should remain at the first column, though.
This plugin modifies some indent commands so that the comment prefix remains in the first column, and the indenting takes place between the comment prefix and the comment text. For that, it uses the comment configuration provided by the buffer's 'comment' option, which is set by most filetype plugins. "