BUG: Dedenting with an indented whitelist character actually increases indent
Whitelist characters don't actually need to be in column 1, so they can have indent, and that may actually be pretty large. So when we dedent (which happens with the prefix temporarily removed), it can happen that the remaining indent is less than l:prefixWidth. In that case, the previous logic just prepended the prefix again, which (with the still existing whitespace) may push the text to the right instead of left.
Instead of asserting on the exact number of spaces and making that match optional, allow any amount of spaces between 0 and that number, so that any remnants of indent-after-prefix are removed.